refactor: change code structure from EAFP to LBYL.

+++ drf/poll/views.py
@@ -1,14 +1,12 @@
-from base64 import urlsafe_b64decode, urlsafe_b64encode
 from datetime import datetime, timedelta
-from django.http.response import JsonResponse, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, get_list_or_404
 from django.contrib.auth.models import User
 
-from rest_framework import generics, views
+from rest_framework import generics
 from rest_framework.request import Request
+from rest_framework.response import Response
 from rest_framework.parsers import JSONParser
 from rest_framework.exceptions import AuthenticationFailed
-from rest_framework.authentication import BaseAuthentication
 
 import jwt
 from jwt.exceptions import DecodeError
@@ -27,51 +25,52 @@ class IndexView(generics.GenericAPIView):
     def get(self, request: Request):
         queryset = self.get_queryset()
         serializer = self.get_serializer(queryset, many=True)
-        return  JsonResponse(serializer.data, safe=False)
+        return  Response(serializer.data)
 
     def post(self, request: Request):
-        try:
-            data = JSONParser().parse(request)
-            serializer = self.get_serializer(data=data)
-            if serializer.is_valid():
-                serializer.save(owner=request.user)
-                return JsonResponse(serializer.data, safe=False)
-            return JsonResponse(serializer.errors, status=400)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(owner=request.user)
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
     
     def patch(self, request: Request):
-        try:
-            queryset = self.get_queryset()
-            data = JSONParser().parse(request)
-            question = get_object_or_404(queryset, pk=data["question_id"])
-            serializer = self.get_serializer(question, data=data, partial=True)
+        queryset = self.get_queryset()
+        question_id = request.data.get("question_id")
+        if question_id != None:
+            try:
+                question_id = int(question_id)
+            except ValueError as e:
+                return Response({"question_id": e.args}, status=400)
+            
+            question = get_object_or_404(queryset, pk=question_id)
+            serializer = self.get_serializer(
+                question, data=request.data, partial=True)
             if serializer.is_valid():
                 serializer.save()
-                return JsonResponse(serializer.data)
-            return JsonResponse(serializer.errors, status=400)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
-    
+                return Response(serializer.data)
+            return Response(serializer.errors, status=400)
+        else:
+            return Response(
+            {"error": f"Missing question_id in request."}, status=400)
+
     def delete(self, request: Request):
-        try:
-            queryset = self.get_queryset()
-            data = JSONParser().parse(request)
-            question_id = int(data['question_id']) 
+        queryset = self.get_queryset()
+        question_id = request.data.get('question_id')
+        if question_id != None:
+            try:
+                question_id = int(question_id)
+            except ValueError as e:
+                return Response({"question_id": e.args}, status=400)
+        
             question = get_object_or_404(queryset, pk=question_id)
             question.delete()
             serializer = QuestionSerializer(question)
-            return JsonResponse(serializer.data, safe=False)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
-    
+            return Response(serializer.data)
+        else:       
+            return Response(
+                    {"error": f"Missing question_id in request."}, status=400)
+
 
 class ChoiceView(generics.GenericAPIView):
     queryset = Choice.objects.all()
@@ -81,53 +80,42 @@ class ChoiceView(generics.GenericAPIView):
         queryset = self.get_queryset()
         choices = get_list_or_404(queryset, question=question_id)
         serializer = self.get_serializer(choices, many=True)
-        return JsonResponse(serializer.data, safe=False)
+        return Response(serializer.data)
     
     def post(self, request: Request, question_id: int):
-        try:
-            data = JSONParser().parse(request)
-            question = Question.objects.get(pk=question_id)
-            serializer = self.get_serializer(data=data)
-            if serializer.is_valid():
-                serializer.save(question=question)
-                return JsonResponse(serializer.data, safe=False)
-            return JsonResponse(serializer.errors, status=400)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
+        question = get_object_or_404(Question, pk=question_id)
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(question=question)
+            return Response(serializer.data)
+        return Response(serializer.errors, status=400)
 
     def patch(self, request: Request, question_id: int):
-        try:
-            queryset = self.get_queryset()
-            data = JSONParser().parse(request)
-            choice = get_object_or_404(queryset, pk=data["choice_id"])
+        queryset = self.get_queryset()
+        choice_id = request.data.get("choice_id")
+        if choice_id != None:
+            try:
+                choice_id = int(choice_id)
+            except ValueError as e:
+                return Response({"choice_id": e.args}, status=400)
+            
+            choice = get_object_or_404(queryset, pk=choice_id)
             serializer = self.get_serializer(
-                choice, data=data, partial=True)
+                choice, data=request.data, partial=True)
             if serializer.is_valid():
                 serializer.save()
-                return JsonResponse(serializer.data, safe=False)
-            return JsonResponse(serializer.errors, status=400)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
+                return Response(serializer.data)
+            return Response(serializer.errors, status=400)
+        else:
+            return Response(
+                {"error": f"Missing choice_id in request."}, status=400)
     
     def delete(self, request: Request, question_id: int):
-        try:
-            queryset = self.get_queryset()
-            data = JSONParser().parse(request)
-            choice = get_object_or_404(queryset, **data)
-            choice.delete()
-            serializer = self.get_serializer(choice)
-            return JsonResponse(serializer.data, safe=False)
-        except KeyError as e:
-            return JsonResponse(
-                {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
+        queryset = self.get_queryset()
+        choice = get_object_or_404(queryset, **request.data)
+        choice.delete()
+        serializer = self.get_serializer(choice)
+        return Response(serializer.data)
     
 
 class UserList(generics.GenericAPIView):
@@ -137,7 +125,7 @@ class UserList(generics.GenericAPIView):
     def get(self, request: Request):
         queryset = self.get_queryset()
         serializer = self.get_serializer(queryset, many=True)
-        return JsonResponse(serializer.data, safe=False)
+        return Response(serializer.data)
 
     
 class UserAuth(generics.GenericAPIView):
@@ -155,7 +143,8 @@ class UserAuth(generics.GenericAPIView):
                 or len(token_split) < 2):
                 raise AuthenticationFailed('Invalid token.')
         
-            payload = jwt.decode(auth_header_split[1], algorithms='HS256', key=SECRET_KEY)
+            payload = jwt.decode(
+                auth_header_split[1], algorithms='HS256', key=SECRET_KEY)
         
             if datetime.now().timestamp() > payload['exp']:
                 raise AuthenticationFailed('Token expired.')
@@ -164,13 +153,12 @@ class UserAuth(generics.GenericAPIView):
             user = get_object_or_404(queryset, pk=payload['uid'])
             serializer = self.get_serializer(user)
         except KeyError as e:
-            return JsonResponse(
+            return Response(
                 {"error": f"Missing {e.args} in request."}, status=400)
         except (DecodeError, User.DoesNotExist) as e:
             raise AuthenticationFailed(e)
 
-        return JsonResponse(serializer.data, safe=False)
-
+        return Response(serializer.data)
 
 
 class LoginView(generics.GenericAPIView):
@@ -181,7 +169,7 @@ class LoginView(generics.GenericAPIView):
         try:
             data = JSONParser().parse(request)
             user = get_object_or_404(
-                self.get_queryset(), username=data['username'])
+                self.get_queryset(), username=data['email'])
             header = {
                 'alg': 'HS256',
                 'typ': 'jwt'
@@ -195,14 +183,14 @@ class LoginView(generics.GenericAPIView):
                 payload=payload, headers=header, 
                 algorithm='HS256', key=SECRET_KEY)
             
-            response = JsonResponse({'jwt': token})
+            response = Response({'jwt': token})
             response.set_cookie(
                 key='jwt', value=token, httponly=True, 
                 expires=timedelta(minutes=60), secure=True)
             return response
         
         except KeyError as e:
-            return JsonResponse(
+            return Response(
                 {"error": f"Missing {e.args} in request."}, status=400)
-        except (TypeError, Question.DoesNotExist) as e:
-            return HttpResponseBadRequest(content=e)
+        except (ValueError, Question.DoesNotExist) as e:
+            return Response({"error": e.args}, status=400)
